requirements.txt
fastapi
uvicorn
gdown
jinja2
numpy
tensorflow
pillow
keras
python-multipart